fix: package update use scikit-learn instead of sklearn; deprecate np.int; a simple setup.py

+++ packages/labelSys/labelSys/utils/labelReaderV2.py
@@ -235,7 +235,7 @@ def readDataSlice(
     elif len(img_ori.shape) == 3:
         im_channel = img_ori.shape[2]
     new_im_size = (
-        (np.array(img_ori.shape[:2]) * magnification).astype(np.int).tolist()
+        (np.array(img_ori.shape[:2]) * magnification).astype(np.int16).tolist()
     )  # H, W
     if USE_SKIMAGE:
         if im_channel == 1:
@@ -308,7 +308,7 @@ def _readOneLabel(ori_im_size, data, magnification, line_thickness):
     - data: data from one label
     return one mask
     """
-    new_im_size = (np.array(ori_im_size) * magnification).astype(np.int)
+    new_im_size = (np.array(ori_im_size) * magnification).astype(np.int16)
     mask = np.zeros(new_im_size[:2], np.uint8)
     cnts = []
     for d in data:
